+++ cash/src/test/java/com/back/cash/CancelPaymentUseCaseTest.java
@@ -1,7 +1,7 @@
 package com.back.cash;
 
 import com.back.cash.adapter.out.PaymentRepository;
-import com.back.cash.app.CancelPaymentUseCase;
+import com.back.cash.app.usecase.CancelPaymentUseCase;
 import com.back.cash.app.CashLogSupport;
 import com.back.cash.app.WalletSupport;
 import com.back.cash.domain.Payment;
